changed the field on the dir_entry from enum FileType to int

+++ include/BlockManagementController.h
@@ -5,9 +5,8 @@
 #ifndef FILESYSTEM_BLOCKMANAGEMENTCONTROLLER_H
 #define FILESYSTEM_BLOCKMANAGEMENTCONTROLLER_H
 
-#include <fsLow.h>
 
-uint64_t ** findSpaceAndWriteToDisk(char *buffer);
+int ** findSpaceAndWriteToDisk(char *buffer);
 
 
 #endif //FILESYSTEM_BLOCKMANAGEMENTCONTROLLER_H
